move meta_train to own file and add some functions needed by meta_training
