+++ util.py
@@ -56,6 +56,19 @@ def make_dataloader(input_ids, labels, attention_masks, batch_size=16, shuffle=T
                             )
     return dataloader
 
+def split_episode(batch, ratio=.75):
+    batch_size = batch[0].size()[0]
+    support_size = round(batch_size*ratio)
+
+    support_idx_pos = np.random.choice(np.arange(0,batch_size,2), np.ceil(support_size/2), replace=False)
+    support_idx_neg = np.random.choice(np.arange(0,batch_size,2), np.floor(support_size/2), replace=False)
+    support_idx = np.sort(np.hstack(support_idx_pos,support_idx_neg))
+    query_idx = np.setdiff1d(np.arange(batch_size), support_idx)
+
+    support_set = (batch[0][support_idx], batch[1][support_idx], batch[2][support_idx])
+    query_set = (batch[0][query_idx], batch[1][query_idx], batch[2][query_idx])
+    return support_set, query_set
+
 
 def get_model():
     """
